Share SubReadStream across System.IO.Compression and System.Formats.Tar
Both ZipArchive and Tar used their own internal SubReadStream classes that exposed a bounded view over a 'super' stream. Tar additionally had a SeekableSubReadStream subclass to support seeking. The two implementations have converged in functionality, so consolidate them into a single sealed class at src/libraries/Common/src/System/IO/SubReadStream.cs that supports both seekable and unseekable super streams.

The unified class also keeps Tar's AdvanceToEnd/AdvanceToEndAsync APIs (no-ops for Zip since Zip never calls them) so TarReader can continue to discard unread data segments.
