+++ src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarHeader.Read.cs
@@ -230,7 +230,8 @@ namespace System.Formats.Tar
         // - Metadata typeflag entries (Extended Attributes and Global Extended Attributes in PAX, LongLink and LongPath in GNU)
         //   will get all the data section read and the stream pointer positioned at the beginning of the next header.
         // - Block, Character, Directory, Fifo, HardLink and SymbolicLink typeflag entries have no data section so the archive stream pointer will be positioned at the beginning of the next header.
-        // - All other typeflag entries with a data section will generate a stream wrapping the data section: SeekableSubReadStream for seekable archive streams, and SubReadStream for unseekable archive streams.
+                    // - All other typeflag entries with a data section will generate a SubReadStream wrapping the data section.
+                    //   When the archive stream is seekable, the SubReadStream is seekable too.
         internal void ProcessDataBlock(Stream archiveStream, bool copyData)
         {
             bool skipBlockAlignmentPadding = true;
@@ -276,15 +277,18 @@ namespace System.Formats.Tar
                         _gnuSparseDataStream = new GnuSparseStream(_dataStream, _gnuSparseRealSize);
                     }
 
-                    if (_dataStream is SeekableSubReadStream)
+                    if (_dataStream is SubReadStream)
                     {
-                        TarHelpers.AdvanceStream(archiveStream, _size);
-                    }
-                    else if (_dataStream is SubReadStream)
-                    {
-                        // This stream gives the user the chance to optionally read the data section
-                        // when the underlying archive stream is unseekable
-                        skipBlockAlignmentPadding = false;
+                        if (archiveStream.CanSeek)
+                        {
+                            TarHelpers.AdvanceStream(archiveStream, _size);
+                        }
+                        else
+                        {
+                            // This stream gives the user the chance to optionally read the data section
+                            // when the underlying archive stream is unseekable
+                            skipBlockAlignmentPadding = false;
+                        }
                     }
 
                     break;
@@ -344,15 +348,18 @@ namespace System.Formats.Tar
                         _gnuSparseDataStream = new GnuSparseStream(_dataStream, _gnuSparseRealSize);
                     }
 
-                    if (_dataStream is SeekableSubReadStream)
+                    if (_dataStream is SubReadStream)
                     {
-                        await TarHelpers.AdvanceStreamAsync(archiveStream, _size, cancellationToken).ConfigureAwait(false);
-                    }
-                    else if (_dataStream is SubReadStream)
-                    {
-                        // This stream gives the user the chance to optionally read the data section
-                        // when the underlying archive stream is unseekable
-                        skipBlockAlignmentPadding = false;
+                        if (archiveStream.CanSeek)
+                        {
+                            await TarHelpers.AdvanceStreamAsync(archiveStream, _size, cancellationToken).ConfigureAwait(false);
+                        }
+                        else
+                        {
+                            // This stream gives the user the chance to optionally read the data section
+                            // when the underlying archive stream is unseekable
+                            skipBlockAlignmentPadding = false;
+                        }
                     }
 
                     break;
@@ -392,9 +399,7 @@ namespace System.Formats.Tar
                 return copiedData;
             }
 
-            return archiveStream.CanSeek
-                ? new SeekableSubReadStream(archiveStream, archiveStream.Position, _size)
-                : new SubReadStream(archiveStream, 0, _size);
+            return new SubReadStream(archiveStream, archiveStream.CanSeek ? archiveStream.Position : 0, _size);
         }
 
         // Asynchronously returns a stream that represents the data section of the current header.
@@ -419,9 +424,7 @@ namespace System.Formats.Tar
                 return copiedData;
             }
 
-            return archiveStream.CanSeek
-                ? new SeekableSubReadStream(archiveStream, archiveStream.Position, size)
-                : new SubReadStream(archiveStream, 0, size);
+            return new SubReadStream(archiveStream, archiveStream.CanSeek ? archiveStream.Position : 0, size);
         }
 
         // Attempts to read the fields shared by all formats and stores them in their expected data type.
