+++ src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarReader.cs
@@ -214,7 +214,7 @@ namespace System.Formats.Tar
                 // When working with seekable streams, every time we return an entry, we avoid advancing the pointer beyond the data section
                 // This is so the user can read the data if desired. But if the data was not read by the user, we need to advance the pointer
                 // here until it's located at the beginning of the next entry header.
-                // This should only be done if the previous entry came from a TarReader and it still had its original SubReadStream or SeekableSubReadStream.
+                // This should only be done if the previous entry came from a TarReader and it still had its original SubReadStream.
 
                 if (_previouslyReadEntry._header._dataStream is not SubReadStream dataStream)
                 {
@@ -247,7 +247,7 @@ namespace System.Formats.Tar
                 // When working with seekable streams, every time we return an entry, we avoid advancing the pointer beyond the data section
                 // This is so the user can read the data if desired. But if the data was not read by the user, we need to advance the pointer
                 // here until it's located at the beginning of the next entry header.
-                // This should only be done if the previous entry came from a TarReader and it still had its original SubReadStream or SeekableSubReadStream.
+                // This should only be done if the previous entry came from a TarReader and it still had its original SubReadStream.
 
                 if (_previouslyReadEntry._header._dataStream is not SubReadStream dataStream)
                 {
