+++ src/libraries/System.IO.Compression/src/Resources/Strings.resx
@@ -371,6 +371,15 @@
   <data name="IO_SeekBeforeBegin" xml:space="preserve">
     <value>An attempt was made to move the position before the beginning of the stream.</value>
   </data>
+  <data name="IO_NotSupported_UnreadableStream" xml:space="preserve">
+    <value>Stream does not support reading.</value>
+  </data>
+  <data name="IO_NotSupported_UnseekableStream" xml:space="preserve">
+    <value>Stream does not support seeking.</value>
+  </data>
+  <data name="IO_NotSupported_UnwritableStream" xml:space="preserve">
+    <value>Stream does not support writing.</value>
+  </data>
   <data name="CrcMismatch" xml:space="preserve">
     <value>The CRC32 checksum of the extracted data does not match the expected value from the archive.</value>
   </data>
